# app/Resources/api_doc.json
{
    "swagger": "2.0",
    "info": {
        "title": "Page API documentation",
        "version": "1.0.0"
    },
    "host": "rochedor.fr",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/news": {
            "get": {
                "summary": "Get requested locale news'list",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "The requested locale",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested news"
                    }
                }
            },
            "post": {
                "summary": "Add a new news",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "intro": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "url": {
                                    "type": "string"
                                },
                                "start": {
                                    "type": "string"
                                },
                                "stop": {
                                    "type": "string"
                                },
                                "locale": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The created news"
                    }
                }
            }
        },
        "/news/{id}/{version}": {
            "get": {
                "summary": "Get a news",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The news id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "description": "The news version",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested news"
                    },
                    "404": {
                        "description": "News not found"
                    }
                }
            }
        },
        "/news/id": {
            "delete": {
                "summary": "Delete requested news",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The news is deleted"
                    },
                    "404": {
                        "description": "news not found"
                    }
                }
            }
        },
        "/news/{id}": {
            "put": {
                "summary": "Edit requested news",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "intro": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "url": {
                                    "type": "string"
                                },
                                "start": {
                                    "type": "string"
                                },
                                "stop": {
                                    "type": "string"
                                },
                                "locale": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the selected news",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": "News not found"
                    }
                }
            }
        },
        "/news/{id}/versions": {
            "get": {
                "summary": "Return all log entries for the selected news",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the news ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested logs of your selected news"
                    },
                    "404": {
                        "description": "News not found"
                    }
                }
            }
        },
...
    },
...
}